refactor(enrichment): split index.ts into data-sources + synthesis (#724)
`src/lib/enrichment/index.ts` was at the absolute ESLint `max-lines: 500` cap
(564 raw lines / ~500 effective with skipBlankLines + skipComments). Any
addition to a wrapper would have breached the cap, blocking further work.

Mirrors how `enrichment-advanced.ts` (LinkedIn + Intelligence Brief +
Outreach) was already split off. The 13 `try*` wrappers now live across
four files; `index.ts` is a thin barrel + `runSingleModule` admin runner.

Layout:
  - data-sources.ts (397 lines) — tryPlaces, tryWebsite, tryOutscraper,
    tryAcc, tryRoc, tryReviewAnalysis, tryCompetitors, tryNews
    (plus the oscOptional / buildOutscraperContent helpers)
  - synthesis.ts (112 lines) — tryDeepWebsite, tryReviewSynthesis
  - enrichment-advanced.ts (144 lines, unchanged) — tryLinkedIn,
    tryIntelligenceBrief, tryOutreach
  - index.ts (115 lines) — public barrel re-exporting all 13 wrappers +
    types + runSingleModule

Public import surface is unchanged. The sole external callsite
(`src/pages/api/admin/entities/[id]/enrichment/[module]/retry.ts`) and
`src/lib/enrichment/workflow.ts`'s 13 wrapper imports both continue to
resolve via the barrel.

`tests/enrichment-pipeline.test.ts` source-reader updated to concatenate
all four files when checking the 13 `export async function try*`
contracts; `tests/outreach-vertical-aware.test.ts` already only needs
index.ts + enrichment-advanced.ts (where tryOutreach lives) and is
unchanged.

Verification:
  npm run lint       # 0 errors, only pre-existing warnings
  npm run typecheck  # 0 errors
  npm run test       # 1736 passed, 2 skipped
  npm run build      # clean
